refactor(frontend): lift user state to App component
- Move user fetching logic from Dashboard to App component
- Pass users and loading state as props to child components
- Remove duplicate API calls for user data
- Update Dashboard and PromptLogs components to use props
- Remove axios import from Dashboard component

This change improves application performance by eliminating duplicate
API calls and centralizes user state management in the App component.
