+++ frontend/src/App.js
@@ -1,10 +1,31 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
+import axios from 'axios';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 import Dashboard from './components/Dashboard';
 import PromptLogs from './components/PromptLogs';
 import './App.css';
 
+
 function App() {
+  const [users, setUsers] = useState([]);
+  const [loadingUsers, setLoadingUsers] = useState(false);
+
+  useEffect(() => {
+    const fetchUsers = async () => {
+      setLoadingUsers(true);
+      try {
+        const response = await axios.get(`${process.env.REACT_APP_API_URL}/users`);
+        setUsers(response.data);
+      } catch (error) {
+        console.error('Error fetching users:', error);
+      } finally {
+        setLoadingUsers(false);
+      }
+    };
+
+    fetchUsers();
+  }, []);
+
   return (
     <Router>
       <div className="min-h-screen bg-gray-100">
@@ -24,8 +45,8 @@ function App() {
         
         <div className="container mx-auto px-4 py-8">
           <Routes>
-            <Route path="/" element={<Dashboard />} />
-            <Route path="/prompt-logs" element={<PromptLogs />} />
+            <Route path="/" element={<Dashboard users={users} loadingUsers={loadingUsers}/>} />
+            <Route path="/prompt-logs" element={<PromptLogs  users={users} loadingUsers={loadingUsers}/>} />
           </Routes>
         </div>
